tests/submission/W3C/domhtml/HTMLSelectElement18.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - SELECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-select-element.html#dom-select-add" rel="help">
<link rel="author" title="Curt Arnold">
</head>
<body>
<form action="./files/getData.pl" id="form1" method="post">
<p>
<select dir="ltr" id="selectId" multiple="multiple" name="select1" size="1" tabindex="7">
<option selected="selected" value="EMP1">EMP10001</option>
<option>EMP10002</option>
<option>EMP10003</option>
<option>EMP10004</option>
<option>EMP10005</option>
</select>
</p>
</form>
<p>
<select name="select2">
<option>EMP20001</option>
<option>EMP20002</option>
<option>EMP20003</option>
<option>EMP20004</option>
<option>EMP20005</option>
</select>
</p>
<p>
<select disabled="disabled" name="select3" tabindex="1">
<option>EMP30001</option>
<option>EMP30002</option>
<option>EMP30003</option>
<option>EMP30004</option>
<option>EMP30005</option>
</select>
</p>








<script>
var domtest = async_test("Add a new option at the end of an select using HTMLSelectElement.add.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var doc;
      var optLength;
      var selected;
      var newOpt;
      var newOptText;
      var opt;
      var optText;
      var optValue;
      var retNode;
      var nullNode = null;

      doc = window.document;
      nodeList = doc.getElementsByTagName("select");
      assert_equals(nodeList.length,3);
testNode = nodeList.item(0);
      newOpt = doc.createElement("option");
      newOptText = doc.createTextNode("EMP31415");
      retNode = newOpt.appendChild(newOptText);
      testNode.add(newOpt,nullNode);
      optLength = testNode.length;

      assert_equals(optLength,6);
       selected = testNode.selectedIndex;

      assert_equals(selected,0);
       opt = testNode.lastChild;

      optText = opt.firstChild;

      optValue = optText.nodeValue;

      assert_equals(optValue,"EMP31415");
       
    domtest.done();
}));

</script>
<p>Add a new option at the end of an select using HTMLSelectElement.add.</p>
<div id="log"></div>
</body></html>
